Reference

TensorMul.broadlike — Method

broadlike(A, B...)

Reshapes (broadcasts) A into the size of A .+ B .+ ..., without doing the sum.

TensorMul.tensordot — Method

tensordot(X, W, Y)

X*W*Y, contracting all dimensions of W with the corresponding first dimensions of X and Y, and matching the remaining last dimensions of X to the remaining last dimensions of Y.

For example, C[b] = sum(X[i,j,b] * W[i,j,μ,ν] * Y[μ,ν,b]).

TensorMul.tensormul_ff — Method

tensormul_ff(A, B, Val(dims))

A*B contracting first dims dimensions of A with first dims dimensions of B (which must match).

TensorMul.tensormul_fl — Method

tensormul_fl(A, B, Val(dims))

A*B contracting first dims dimensions of A with last dims dimensions of B.

TensorMul.tensormul_lf — Method

tensormul_lf(A, B, Val(dims))

A*B contracting last dims dimensions of A with first dims dimensions of B.

TensorMul.tensormul_ll — Method

tensormul_ll(A, B, Val(dims))

A*B contracting last dims dimensions of A with last dims dimensions of B.